FM-24 FM-7 : Add Email Verification Checking
User is not prevented from logging in if their email is not verified.
If not, they are redirected to a new screen. After they verify,
they can tap the link to try again. No re-login is necessary.

Prevented stack overflow with refreshToken refactor. Decided to add error handling which resolved FM-7.