FM-20 : Add Local Storage
Integrated a NoSQL encrypted database to cache event and ticket information for offline use.
If force refresh is set to false, the offline values are used.
This allows the device to load data if it is offline.

Removed the venue storage since it is not used.
Removed old dashboard file that was not called.
Unit tests require updating.