FM-4 : Corrected Tests
Refactored MyPassTab to accept constructors for auth utils and ticket provider. This allows for it to be directly tested and not build home.

Nasty issue was resolved with building a ChangeNotifier. We were incorrectly using the builder. I found a comment buried deep in the code that said you must use the value method call if the provider is already build. I only assume this was missed but an issue until we did the refactoring.

See:
- DON'T reuse an existing ChangeNotifier using the default constructor.

MyChangeNotifier variable;
ChangeNotifierProvider(
  builder: (_) => variable,
  child: ...
)